[GOVCMS-4390]: Chunk govcms-deploy.
This should provide small testable pieces of the govcms-deploy script which can be
ordered in the .lagoon.yml file. This will allow PaaS users to order their deployment
process but also use supported scripts.

It will allow us to more quickly reorganise the scripts if we need to change the
deployment process for a particular release.